Set all animations
