+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedCambuci.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95ac21b2deed62d4ab7bbd34cda30aaa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
