+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedCascaDura.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d15e44abf1867674aa6de5b5438e0cc3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
