+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedFertilizer.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 728fd4567866a8042a63bb3174c28046
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
