+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedPesticide.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19611585d14e14646a77729a921bec10
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
