+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedSpawn.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d65309d52713bcb48953129ac3ad6d94
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
