+++ Assets/Animation/PowerUpPrizeDraw/clip_SelectedWasp.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5254191283e34c4eb1d2cb7d1612de9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
