+++ Assets/Prefabs/PowerUp/pw_Arbusto.prefab
@@ -46,11 +46,11 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   PowerUp: {fileID: 7921277358032559249}
-  PercentagePosition1: 50
-  PercentagePosition2: 50
-  PercentagePosition3: 50
-  PercentagePosition4: 50
-  PercentagePosition5: 0
+  PercentagePosition1: 15
+  PercentagePosition2: 5
+  PercentagePosition3: 5
+  PercentagePosition4: 10
+  PercentagePosition5: 5
   PercentagePosition6: 0
   PercentagePosition7: 0
   PercentagePosition8: 0
@@ -98,6 +98,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e4ac51955c9ffaf4aacb69d34517160e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _id: 1
-  _name: Arbusto
+  _id: 7
+  _name: Arbusto Cortante
   _thumbnail: {fileID: 2800000, guid: fb413acb35eb19d4c9ff4f3a13eb2982, type: 3}
