--- Assets/Scripts/Entities/PowerUp.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e4ac51955c9ffaf4aacb69d34517160e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
